High Scores: use the instance method to access the @scores variable, rather than accessing the variable itself. Also use the "last" method for the array, rather than [-1] - easier to read.
-- Full mentor feedback:
Hey there! Nice work on this one. Let's look at a couple of changes to tighten this up a bit.

You can replace your scores method with an attr_reader, which is a macro for declaring getters. It expands to exactly what
you have here but with a lot less syntactic noise. All you do is put attr_reader :scores in the class definition and things
will just work. They'll work anywhere in the class definition although it is customary to put them up at the top.

In Ruby it is typical to leverage the other instance methods available to you in an object. In this case you could use the
scores getter method throughout as opposed to referring to the underlying @ variable. In a real project this would matter
for extending classes that are overriding behavior, and it also makes it easier to change behavior if you need to.

Instead of using the bracket notation to get the last element from the array you could call .last instead. It does exactly
the same thing but is a little more readable.

Give those changes a try and see how it looks!