Scrabble Score: mentor tip - use sum(&tiles) as a shortcut to a block which takes a letter and looks up its value in the tiles hash.
Where we do sum the letters, we might also do letters.sum(:tiles) or something very close to that. It is a shortcut way to
work with hashes as the subject of iteration. Each letter being iterated would be given to the hash named there in the argument as a symbol.